airflow-ga4/requirements.txt
requests
google-api-python-client
oauth2client